# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Fluxend is backend as-a-service platform that allows you to build, deploy, and scale applications without managing infrastructure.",
        "title": "Fluxend API",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/fluxend/fluxend",
            "email": "hello@fluxend.app"
        },
        "version": "1.0"
    },
    "paths": {
        "/admin/logs": {
            "get": {
                "description": "Get all logs",
                "tags": [
                    "Logs"
                ],
                "summary": "List logs",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by user UUID",
                        "name": "userUuid",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by HTTP method",
                        "name": "method",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by endpoint",
                        "name": "endpoint",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by IP address",
                        "name": "ipAddress",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Field to sort by",
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Sort order (asc or desc)",
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of files",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "allOf": [
                                            {
                                                "$ref": "#/components/schemas/response.Response"
                                            },
                                            {
                                                "type": "object",
                                                "properties": {
                                                    "content": {
                                                        "type": "array",
                                                        "items": {
                                                            "$ref": "#/components/schemas/logging.Response"
                                                        }
                                                    }
                                                }
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/tables/{tableUUID}/indexes": {
            "get": {
                "description": "Retrieve a list of indexes for a given table.",
                "tags": [
                    "Indexes"
                ],
                "summary": "List indexes",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Table UUID",
                        "name": "tableUUID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of indexes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response.Response"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "content": {
                                                    "type": "array",
                                                    "items": {}
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "description": "Add an index to a specified table within a project.",
                "tags": [
                    "Indexes"
                ],
                "summary": "Create index",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Table UUID",
                        "name": "tableUUID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/database.CreateIndexRequest"
                            }
                        }
                    },
                    "description": "Index details JSON",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Index created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response.Response"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "content": {}
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/tables/{tableUUID}/indexes/{indexName}": {
            "get": {
                "description": "Retrieve details for a specific index in a table.",
                "tags": [
                    "Indexes"
                ],
                "summary": "Retrieve index",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Table UUID",
                        "name": "tableUUID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Index Name",
                        "name": "index_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Index details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response.Response"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "content": {}
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Index not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "description": "Remove an existing index from a given table.",
                "tags": [
                    "Indexes"
                ],
                "summary": "Delete index",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Table UUID",
                        "name": "tableUUID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Index Name",
                        "name": "index_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Index deleted successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Index not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "//api.fluxend.app"
        }
    ],
    "components": {
...
        "schemas": {
...
            "database.CreateIndexRequest": {
                "type": "object",
                "properties": {
                    "columns": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "is_unique": {
                        "type": "boolean"
                    },
                    "name": {
                        "type": "string"
                    },
                    "projectUUID": {
                        "type": "string"
                    }
                }
            },
...
            "logging.Response": {
                "type": "object",
                "properties": {
                    "body": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "endpoint": {
                        "type": "string"
                    },
                    "ipAddress": {
                        "type": "string"
                    },
                    "method": {
                        "type": "string"
                    },
                    "params": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "userAgent": {
                        "type": "string"
                    },
                    "userUuid": {
                        "type": "string"
                    },
                    "uuid": {
                        "type": "string"
                    }
                }
            },
...
            "response.Response": {
                "type": "object",
                "properties": {
                    "content": {},
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
        }
    }
}